requirements.txt
discord~=1.7.0
python-dotenv~=0.15.0
PyYAML~=5.4.0
aiopg~=1.3.0
SQLAlchemy~=1.3.24